Remove unnecessary pa api module and move error.rs, stream.rs and types.rs modules to top level for a more rustesque module layout. Update examples accordingly.
