+++ src/stream.rs
